Refactored packages management, implemented documentos endpoint

+++ backend/src/main/resources/db/migration/V1__tabelas_iniciais.sql
@@ -12,7 +12,7 @@ CREATE TABLE usuarios (
                           nome VARCHAR(255) NOT NULL,
                           email VARCHAR(255) NOT NULL UNIQUE,
                           senha VARCHAR(255) NOT NULL,
-                          setor_id INTEGER REFERENCES setores(id),
+                          setor_id BIGINT REFERENCES setores(id),
                           perfil VARCHAR(50) NOT NULL, -- Ex: 'ADMIN', 'GESTOR', 'USUARIO'
                           created_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP
 );
@@ -24,8 +24,8 @@ CREATE TABLE documentos (
                             descricao TEXT,
                             tipo VARCHAR(100) NOT NULL, -- Ex: 'Edital', 'Relatório', 'Ata'
                             caminho_arquivo VARCHAR(255) NOT NULL,
-                            setor_id INTEGER NOT NULL REFERENCES setores(id),
-                            usuario_id INTEGER NOT NULL REFERENCES usuarios(id),
+                            setor_id BIGINT NOT NULL REFERENCES setores(id),
+                            usuario_id BIGINT NOT NULL REFERENCES usuarios(id),
                             created_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP,
                             updated_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP
 );
@@ -33,11 +33,11 @@ CREATE TABLE documentos (
 -- Tabela para controle de versionamento dos documentos
 CREATE TABLE versoes_documento (
                                    id BIGSERIAL PRIMARY KEY,
-                                   documento_id INTEGER NOT NULL REFERENCES documentos(id) ON DELETE CASCADE,
-                                   versao INTEGER NOT NULL,
+                                   documento_id BIGINT NOT NULL REFERENCES documentos(id) ON DELETE CASCADE,
+                                   versao BIGINT NOT NULL,
                                    caminho_arquivo VARCHAR(255) NOT NULL,
                                    motivo_alteracao TEXT,
-                                   usuario_id INTEGER NOT NULL REFERENCES usuarios(id),
+                                   usuario_id BIGINT NOT NULL REFERENCES usuarios(id),
                                    created_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP,
                                    UNIQUE(documento_id, versao)
 );
@@ -47,8 +47,8 @@ CREATE TABLE logs (
                       id BIGSERIAL PRIMARY KEY,
                       acao VARCHAR(255) NOT NULL, -- Ex: 'UPLOAD', 'DOWNLOAD', 'LOGIN'
                       descricao TEXT,
-                      usuario_id INTEGER REFERENCES usuarios(id),
-                      documento_id INTEGER REFERENCES documentos(id),
+                      usuario_id BIGINT REFERENCES usuarios(id),
+                      documento_id BIGINT REFERENCES documentos(id),
                       "timestamp" TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP
 );
